ditto

+++ Globals/Classes/Time/Clock.cs
@@ -28,7 +28,7 @@ namespace Globals.Classes.Time
         public static TimeKeeper TimeKeeper;
 
         //Increases by 1 every time minutes increase, never decreases
-        public static int TotalTime;
+        public static float TotalTime;
 
         public static bool Paused;
